Verify document full-linkage in Unmarshal
Unmarshal is typically used for deserializing JSON:API documents found
in request bodies, which may have relationships, but which are not
compound documents (with an `included` section). However, users of this
library may want to deserialize JSON:API-compliant server responses in
full. To ensure that standards are being met for that use case, compound
document full-linkage must be asserted during Unmarshal.