Consider nil `*Link`s as per se valid
Nothing in the JSON:API spec requires links (usually e.g. it’s either
links or data), and nothing requires that if one relation has a related
resource link, all must.

This allows `LinkRelation` to return nil for related resources which may
not have exposed linkage routes in an API.